+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `#[builder(crate_module_path = ...)]` for overcoming cases where the derive
+  macro is used in another crate's macro (see issue #109)
 
 ## 0.15.2 - 2023-08-03
 ### Fixed